支持PASEF模式,压缩率仅和vendor文件差不多

+++ DomainsCore/Scan/TimsData.cs
@@ -2,11 +2,11 @@
 
 public class TimsData
 { 
-    public float mobility;
+    public double mobility;
     public double mz;
     public double intensity;
 
-    public TimsData(float mobility, double mz, double intensity)
+    public TimsData(double mobility, double mz, double intensity)
     {
         this.mobility = mobility;
         this.mz = mz;
